backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
supabase==2.0.3
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.1
starlette==0.36.3
email-validator==2.1.0.post1
